This fix an issue with the <testsuites> element that was not taken into account by the XML configuration class. The suite was empty, so this new feature was simply not working. Not it uses the testsuites defined in the XML document.
Since I am not familiar with the PHPUnit API, I am wondering if we are also missing a few calls such as:

getLoggingConfiguration()
getGroupConfiguration()
getFilterConfiguration()
getPHPConfiguration()

So that all the sections of the XML configuration file get properly added to the "run()" call.

With this fix, the testsuites section of the XML configuration file will be properly working with VisualPHPUnit, but other sections of the configuration file may not be taken into account.